Changing PK type cleanup
Change PK data type from big unsigned int to SERIAL for cose_user, general_care_plan and care_needs

+++ database/migrations/2025_03_12_050540_create_health_histories_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('health_histories', function (Blueprint $table) {
             $table->increments('health_history_id');
-            $table->unsignedBigInteger('general_care_plan_id')->after('health_history_id');
+            $table->integer('general_care_plan_id')->after('health_history_id');
             $table->integer('history_category_id')->after('general_care_plan_id');
             $table->text('history_description')->after('history_category_id');
             $table->timestamps();
